after login and save alerts for approvals

+++ SAP/util/GlobalVar.cs
@@ -15,7 +15,7 @@ namespace SAP.util
         public static bool isReady = false;
         public static String cardCode;
         public static DateTime datetime;
-        public static int usuarioId;
+        public static Int16 usuarioId;
 
         public static void inicializarEmpresa()
         {
